+++ NEWS.md
@@ -8,6 +8,8 @@
 * Better handling of larger datasets when running `tabnet_explain`.
 * Batch predictions to avoid OOM error.
 * Add tabnet_pretrain() for unsupervised pretraining (@cregouby, #29)
+* Add autoplot() of model loss among epochs (@cregouby, #36)
+* Added a `config` argument to fit/pretrain functions so one can pass a pre-made config list. (#42)
 * Add `mask_type` configuration option with `entmax` additional to `sparsemax` (@cmcmaster1, #48)
 
 # tabnet 0.1.0